+++ lib/ReactViews/Search/SearchBoxAndResults.jsx
@@ -181,16 +181,17 @@ export class SearchBoxAndResultsRaw extends React.Component {
               {/* search {searchterm} in data catalog */}
               {/* ~TODO: Put this back once we add a MobX DataCatalogSearch Provider~ */}
               {/* TODO2: Implement a more generic MobX DataCatalogSearch */}
-              {searchState.catalogSearchProvider && (
-                <Box column paddedRatio={2}>
-                  <SearchInDataCatalog
-                    viewState={viewState}
-                    handleClick={() => {
-                      this.toggleShowLocationSearchResults(false);
-                    }}
-                  />
-                </Box>
-              )}
+              {this.props.terria.searchBarModel.showSearchInCatalog &&
+                searchState.catalogSearchProvider && (
+                  <Box column paddedRatio={2}>
+                    <SearchInDataCatalog
+                      viewState={viewState}
+                      handleClick={() => {
+                        this.toggleShowLocationSearchResults(false);
+                      }}
+                    />
+                  </Box>
+                )}
               <Box
                 column
                 css={`