+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 #### next release (8.7.9)
 
+- Add "searchBarConfig.showSearchInCatalog" to configParameters so that the link in location search results can be disabled.
 - Properly initialize react ref in functional components
 - Add NominatimSearchProvider.
 - [The next improvement]